feat: configure typescript eslint and prettier
