+++ src/http_server/index.ts
@@ -1,0 +1,18 @@
+import * as fs from 'fs';
+import * as http from 'http';
+import * as path from 'path';
+
+export const httpServer = http.createServer(function (req, res) {
+  const __dirname = path.resolve(path.dirname(''));
+  const file_path =
+    __dirname + (req.url === '/' ? '/front/index.html' : '/front' + req.url);
+  fs.readFile(file_path, function (err, data) {
+    if (err) {
+      res.writeHead(404);
+      res.end(JSON.stringify(err));
+      return;
+    }
+    res.writeHead(200);
+    res.end(data);
+  });
+});
